--- src/infra/app.controller.ts
@@ -1,28 +1,0 @@
-import { Body, Controller, Get, Post } from '@nestjs/common';
-import { randomUUID } from 'node:crypto';
-import { CreateNotificationBody } from './create-notification-body';
-import { PrismaService } from './prisma.service';
-
-@Controller('notifications')
-export class AppController {
-  constructor(private readonly prisma: PrismaService) {}
-
-  @Get()
-  list() {
-    return this.prisma.notification.findMany();
-  }
-
-  @Post()
-  async create(@Body() body: CreateNotificationBody) {
-    const { recipientId, content, category } = body;
-
-    await this.prisma.notification.create({
-      data: {
-        id: randomUUID(),
-        content,
-        category,
-        recipientId,
-      },
-    });
-  }
-}
