+++ src/infra/database/database.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { PrismaService } from './prisma/prisma.service';
+
+@Module({
+  providers: [PrismaService],
+})
+export class DatabaseModule {}
