--- src/infra/http.module.ts
@@ -1,6 +1,0 @@
-import { Module } from '@nestjs/common';
-
-@Module({
-  imports: [],
-})
-export class HttpModule {}
