+++ src/infra/http/controllers/notifications.controller.ts
@@ -1,0 +1,11 @@
+import { Body, Controller, Post } from '@nestjs/common';
+import { CreateNotificationBody } from '../dtos/create-notification-body';
+
+@Controller('notifications')
+export class NotificationsController {
+  // constructor(private readonly prisma: PrismaService) {}
+  @Post()
+  async create(@Body() body: CreateNotificationBody) {
+    const { recipientId, content, category } = body;
+  }
+}
