+++ src/infra/http/http.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { NotificationsController } from './controllers/notifications.controller';
+
+@Module({
+  imports: [],
+  controllers: [NotificationsController],
+})
+export class HttpModule {}
